+++ src/parser/inline/extension/index.test.ts
@@ -19,8 +19,8 @@ describe('Unit: parser/inline/extension/index', () => {
       assert.deepStrictEqual(inspect(parser('[#\n]')), undefined);
       assert.deepStrictEqual(inspect(parser('[#\\\n]')), undefined);
       assert.deepStrictEqual(inspect(parser('[#\\]')), undefined);
-      assert.deepStrictEqual(inspect(parser('[#[]{}]')), undefined);
-      assert.deepStrictEqual(inspect(parser('[#![]{}]')), undefined);
+      assert.deepStrictEqual(inspect(parser('[#[]{a}]')), undefined);
+      assert.deepStrictEqual(inspect(parser('[#![]{a}]')), undefined);
     });
 
     it('basic', () => {