+++ src/parser/inline/link.test.ts
@@ -39,6 +39,7 @@ describe('Unit: parser/inline/link', () => {
 
     it('invalid', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
+      assert.deepStrictEqual(inspect(parser('[]{}')), undefined);
       assert.deepStrictEqual(inspect(parser('[]{ }')), undefined);
       assert.deepStrictEqual(inspect(parser('[]{  }')), undefined);
       assert.deepStrictEqual(inspect(parser('[]{/  }')), undefined);
@@ -55,8 +56,8 @@ describe('Unit: parser/inline/link', () => {
       assert.deepStrictEqual(inspect(parser('[ ]{ }')), undefined);
       assert.deepStrictEqual(inspect(parser('[[]{}')), undefined);
       assert.deepStrictEqual(inspect(parser('[]]{}')), undefined);
-      assert.deepStrictEqual(inspect(parser('[ a]{}')), undefined);
-      assert.deepStrictEqual(inspect(parser('[ a ]{}')), undefined);
+      assert.deepStrictEqual(inspect(parser('[ a]{b}')), undefined);
+      assert.deepStrictEqual(inspect(parser('[ a ]{b}')), undefined);
       assert.deepStrictEqual(inspect(parser('[a\nb]{ab}')), undefined);
       assert.deepStrictEqual(inspect(parser('[a\\\nb]{ab}')), undefined);
       assert.deepStrictEqual(inspect(parser('[*a\\]{\nb*]{ab}')), undefined);
@@ -67,7 +68,6 @@ describe('Unit: parser/inline/link', () => {
     });
 
     it('basic', () => {
-      assert.deepStrictEqual(inspect(parser('[]{}')), [[`<a href="" rel="noopener">.</a>`], '']);
       assert.deepStrictEqual(inspect(parser('[]{b}')), [['<a href="b" rel="noopener">b</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[]{b }')), [['<a href="b" rel="noopener">b</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[]{ b }')), [['<a href="b" rel="noopener">b</a>'], '']);
@@ -77,10 +77,9 @@ describe('Unit: parser/inline/link', () => {
       assert.deepStrictEqual(inspect(parser('[]{a b}')), [['<a href="a" rel="noopener" class="invalid" data-invalid-type="parameter">a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[]{?b=c+d&\\#}')), [['<a href="?b=c+d&amp;\\#" rel="noopener">?b=c+d&amp;\\#</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[]{?&amp;}')), [['<a href="?&amp;amp;" rel="noopener">?&amp;amp;</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[a]{}')), [['<a href="" rel="noopener">a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[a]{b}')), [['<a href="b" rel="noopener">a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[a]{#}')), [['<a href="#" rel="noopener">a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[a ]{}')), [['<a href="" rel="noopener">a </a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[a ]{b}')), [['<a href="b" rel="noopener">a </a>'], '']);
       assert.deepStrictEqual(inspect(parser('[a b]{c}')), [['<a href="c" rel="noopener">a b</a>'], '']);
       assert.deepStrictEqual(inspect(parser(`[]{?#${encodeURIComponent(':/[]{}<>?#=& ')}}`)), [['<a href="?#%3A%2F%5B%5D%7B%7D%3C%3E%3F%23%3D%26%20" rel="noopener">?#%3A%2F[]{}&lt;&gt;%3F%23%3D%26%20</a>'], '']);
     });