+++ src/parser/inline/media.test.ts
@@ -19,6 +19,7 @@ describe('Unit: parser/inline/media', () => {
 
     it('invalid', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
+      assert.deepStrictEqual(inspect(parser('![]{}')), undefined);
       assert.deepStrictEqual(inspect(parser('![]{ }')), undefined);
       assert.deepStrictEqual(inspect(parser('![]{  }')), undefined);
       assert.deepStrictEqual(inspect(parser('![]{/  }')), undefined);
@@ -30,8 +31,8 @@ describe('Unit: parser/inline/media', () => {
       assert.deepStrictEqual(inspect(parser('![]{a\nb}')), undefined);
       assert.deepStrictEqual(inspect(parser('![]{a\\\nb}')), undefined);
       assert.deepStrictEqual(inspect(parser('![]{{\n}}')), undefined);
-      assert.deepStrictEqual(inspect(parser('![ a]{}')), undefined);
-      assert.deepStrictEqual(inspect(parser('![ a ]{}')), undefined);
+      assert.deepStrictEqual(inspect(parser('![ a]{#}')), undefined);
+      assert.deepStrictEqual(inspect(parser('![ a ]{#}')), undefined);
       assert.deepStrictEqual(inspect(parser('![a\nb]{ab}')), undefined);
       assert.deepStrictEqual(inspect(parser('![a\\\nb]{ab}')), undefined);
       assert.deepStrictEqual(inspect(parser('![*a\\]{\nb*]{ab}')), undefined);
@@ -42,7 +43,6 @@ describe('Unit: parser/inline/media', () => {
     });
 
     it('basic', () => {
-      assert.deepStrictEqual(inspect(parser('![]{}')), [['<img class="media" data-src="" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![]{b}')), [['<img class="media" data-src="b" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![]{b }')), [['<img class="media" data-src="b" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![]{ b }')), [['<img class="media" data-src="b" alt="">'], '']);
@@ -50,9 +50,8 @@ describe('Unit: parser/inline/media', () => {
       assert.deepStrictEqual(inspect(parser('![]{\\ }')), [['<img class="media" data-src="\\" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![]{\\b}')), [['<img class="media" data-src="\\b" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![]{a b}')), [['<img class="media invalid" data-src="a" alt="" data-invalid-type="parameter">'], '']);
-      assert.deepStrictEqual(inspect(parser('![a]{}')), [['<img class="media" data-src="" alt="a">'], '']);
-      assert.deepStrictEqual(inspect(parser('![a ]{}')), [['<img class="media" data-src="" alt="a">'], '']);
-      assert.deepStrictEqual(inspect(parser('![a b]{}')), [['<img class="media" data-src="" alt="a b">'], '']);
+      assert.deepStrictEqual(inspect(parser('![a ]{b}')), [['<img class="media" data-src="b" alt="a">'], '']);
+      assert.deepStrictEqual(inspect(parser('![a b]{c}')), [['<img class="media" data-src="c" alt="a b">'], '']);
     });
 
     it('nest', () => {